pyproject.toml
[tool.poetry]
name = "flash-loan-arbitrage"
version = "0.1.0"
description = "Flash Loan Arbitrage Bot with Python"
authors = ["Your Name <your.email@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
web3 = {version = "^6.11.3", extras = ["websockets"]}
python-dotenv = "^1.0.0"
eth-typing = "^3.5.1"
eth-utils = "^2.3.1"
requests = "^2.31.0"
websockets = "^12.0"
aiohttp = "^3.9.1"
eth-account = "^0.9.0"
eth-hash = {extras = ["pycryptodome"], version = "^0.5.2"}
setuptools = "^69.0.3"
eth-keyfile = "^0.6.1"
cytoolz = "^0.12.2"
eth-rlp = "^0.3.0"
eth-keys = "^0.4.0"
eth-abi = "^4.2.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
start = "src.main:main"
